==> code/mention-classification/round2/04-llm_zeroshot_attribute_annotation.ipynb <==
# TODO:
#  - decide which parts to export to module because they can be recycled

import os
import pandas as pd
import yaml
import json
from pprint import pprint

from llama_index.llms.ollama import Ollama
from llama_index.core.program import LLMTextCompletionProgram 

from pydantic import BaseModel, ValidationError
from typing import List, Literal

## Prepare the data

data_path = '../../../data/annotations/group_mention_categorization/social-group-mention-categorization-round02'
fp = os.path.join(data_path, 'consolidated_annotations.tsv')
df = pd.read_csv(fp, sep='\t')

# # inspect stance annotations
# df.loc[df.attribute=='stance', 'label'].value_counts()
# df.loc[df.label=='Negative', ['text', 'mention']].drop_duplicates()

attributes = ['economic', 'non-economic']
df = df[df.attribute.isin(attributes)]

# inspect distribution of econ attributes
tmp = df[df.attribute==attributes[0]].value_counts(['category', 'label']).reset_index().pivot(index='category', columns='label', values='count')
tmp.fillna(0, inplace=True)
tmp['n'] = tmp[['No', 'Yes']].sum(axis=1)
tmp['prevalence'] = tmp['Yes'] / tmp['n']
tmp.round(3)

# inspect distribution of non-econ attributes
tmp = df[df.attribute==attributes[1]].value_counts(['category', 'label']).reset_index().pivot(index='category', columns='label', values='count')
tmp.fillna(0, inplace=True)
tmp['n'] = tmp[['No', 'Yes']].sum(axis=1)
tmp['prevalence'] = tmp['Yes'] / tmp['n']
tmp.round(3)

tmp = df[df.attribute==attributes[1]].value_counts(['category', 'label']).reset_index().pivot(index='category', columns='label', values='count')
tmp['n'] = tmp[['No', 'Yes']].sum(axis=1)
tmp

### format annotations into format suitable for analyses

# gather labels and pivot wider
exemplars = df.groupby(['mention_id', 'mention', 'attribute']).\
    apply(lambda x: '; '.join(x.category[x.label=='Yes'].tolist()), include_groups=False).\
    reset_index().\
    rename(columns={0: 'labels'}).\
    pivot(index=['mention_id', 'mention'], columns='attribute', values='labels').\
    reset_index()

# discard mentions with disagreeing labels on any of the attributes
has_disagreements = lambda x : (x[attributes].apply(lambda a: a.nunique(), axis=0)>1).any()
disagreements = exemplars.groupby('mention').apply(has_disagreements, include_groups=False)
exemplars = exemplars[~exemplars.mention.isin(disagreements[disagreements].index)]

# drop duplicates
exemplars = exemplars[['mention'] + attributes].drop_duplicates()

# split into list of labels
exemplars.loc[:, attributes] = exemplars[attributes].apply(lambda x: x.str.split('; '))
for a in attributes:
    exemplars.loc[:, a] = exemplars[a].apply(lambda x: [e for e in x if e!=''])

## Prepare the prompt template

attributes_file = '../../../data/annotations/group_mention_categorization/group_attributes_v2.yaml'
with open(attributes_file, 'r') as f:
    ontology = yaml.load(f, Loader=yaml.FullLoader)

scheme = {}
for dim, d in ontology['social_group'].items():
    scheme[d['display_name']] = {
        'other ' + d['display_name'][:-1] if k.startswith('<i>Other') else k:
        '"' + '", "'.join(v) + '", etc.' if v else None
        for k, v in d['attributes'].items()
    }

pprint(scheme['economic attributes'])

# see template examples: https://github.com/run-llama/llama_index/blob/main/llama-index-core/llama_index/core/prompts/default_prompts.py
prompt_template = """\
Act as a multilabel classifier for social group attributes.

Below you will see a text that refers to a social group or a collective of people.

Your task is to classify which {attributes_dim} are explicitly used to define or describe the mentioned group, if any. Choose from the following categories:

{attributes_list}

Return your response as a JSON list of strings, where each string is a category that is used in the text to define or describe the mentioned group.
If you think that the group is not described by any of these {attributes_dim}, return an empty list `[]`.

INPUT: \"\"\"{mention}\"\"\"
RESPONSE:\
"""

def create_attributes_list(atrributes_dict):
    return '\n'.join([f'- **{k}**: {v}' if v else f'- **{k}**' for k, v in atrributes_dict.items()])

prompt_templates = {
    dim:
    prompt_template.format(
        attributes_dim=dim, 
        attributes_list=create_attributes_list(attrs),
        mention='{mention}'
    )
    for dim, attrs in scheme.items()
}

# init LLM
llm = Ollama(
    model='phi4',
    temperature=0.0,
    seed=42,
    json_mode=True
)

# Example
llm.complete('How is the whether today?')

dim = 'economic attributes'

Category = Literal[*list(scheme[dim].keys())]
class Categories(BaseModel):
    """Data model for selected categories (if any)"""
    categories: List[Category]

econ_attributes_classifier = LLMTextCompletionProgram.from_defaults(
    output_cls=Categories,
    prompt_template_str=prompt_templates[dim],
    llm=llm,
    verbose=True,
)

dim = 'non-economic attributes'

Category = Literal[*list(scheme[dim].keys())]
class Categories(BaseModel):
    """Data model for selected categories (if any)"""
    categories: List[Category]

nonecon_attributes_classifier = LLMTextCompletionProgram.from_defaults(
    output_cls=Categories,
    prompt_template_str=prompt_templates[dim],
    llm=llm,
    verbose=True,
)

ex = exemplars.iloc[0]
print('"""'+ex.mention+'"""')
print(' -     econ:', ex['economic'], '(humans);', econ_attributes_classifier(mention=ex.mention).categories, '(llm)')
print(' - non-econ:', ex['non-economic'], '(humans);', nonecon_attributes_classifier(mention=ex.mention).categories, '(llm)')

from tqdm.auto import tqdm

llm_annotations = {a: {} for a in attributes}

for i, ex in tqdm(exemplars.iterrows(), total=len(exemplars)):
    # econ attributes
    try:
        llm_annotations['economic'][i] = econ_attributes_classifier(mention=ex.mention).categories
    except ValidationError:
        pass
    
    # non-economic attributes
    try:
        llm_annotations['non-economic'][i] = nonecon_attributes_classifier(mention=ex.mention).categories
    except ValidationError:
        pass

# TODO: check how to handle validation errors in pydantic 
#       (see https://docs.pydantic.dev/2.9/errors/validation_errors 
#        and llama_index.core.types.BaseOutputParser class)

exemplars = exemplars.\
    join(pd.DataFrame(pd.Series(llm_annotations['economic']), columns=['economic_pred'])).\
    join(pd.DataFrame(pd.Series(llm_annotations['non-economic']), columns=['non-economic_pred']))

### Compute agreement

# one-hot encode annotations
dim = 'economic'
human = exemplars[dim].apply(lambda x: [int(a in x) for a in scheme[dim+' attributes'].keys()])
llm = exemplars[dim+'_pred'].apply(lambda x: [int(a in x) for a in scheme[dim+' attributes'].keys()])

import numpy as np
human = np.stack(human.values)
llm = np.stack(llm.values)

# compute f1 score for multilabel classification
from sklearn.metrics import classification_report

print(classification_report(human, llm, target_names=scheme[dim+' attributes'].keys(), zero_division=0.0))

# one-hot encode annotations
dim = 'non-economic'
human = exemplars[dim].apply(lambda x: [int(a in x) for a in scheme[dim+' attributes'].keys()])
llm = exemplars[dim+'_pred'].apply(lambda x: [int(a in x) for a in scheme[dim+' attributes'].keys()])

import numpy as np
human = np.stack(human.values)
llm = np.stack(llm.values)

# compute f1 score for multilabel classification
from sklearn.metrics import classification_report

print(classification_report(human, llm, target_names=scheme[dim+' attributes'].keys(), zero_division=0.0))

## Export disagreement cases

for dim in attributes:
    exemplars[dim] = exemplars[dim].apply(lambda x: [int(a in x) for a in scheme[dim+' attributes'].keys()])
    exemplars[dim+'_pred'] = exemplars[dim+'_pred'].apply(lambda x: [int(a in x) for a in scheme[dim+' attributes'].keys()])

idxs = np.logical_or(exemplars['economic'] != exemplars['economic_pred'], exemplars['non-economic'] != exemplars['non-economic_pred'])

tmp = exemplars[idxs].copy()
for dim in attributes:
    tmp.loc[:, dim+'_category'] = [list(scheme[dim+' attributes'].keys())]*len(tmp)
# pivot longer
tmp = tmp.melt(id_vars=['mention'], var_name='attribute', value_name='labels')
tmp[['attribute', 'source']] = tmp.attribute.str.split('_', expand=True).replace({None: 'annotators', 'pred': 'llm'})
# pivot wider
tmp = tmp.pivot(index=['mention', 'attribute'], columns='source', values='labels').reset_index()
# explode
tmp = tmp.explode(['category', 'annotators', 'llm', ])

# NOTE: can be subset to ~280 disagreement cases
tmp[tmp.annotators!=tmp.llm]

out = df[['mention_id', 'text', 'mention']].drop_duplicates().merge(tmp[['mention', 'attribute', 'category', 'annotators', 'llm']], how='inner', on='mention')

out.query('annotators!=llm').sort_values(['mention', 'mention_id']).to_clipboard()

data_path = '../../../data/annotations/group_mention_categorization/social-group-mention-categorization-llm-review'
fp = os.path.join(data_path, 'cases.tsv')
out.to_csv(fp, sep='\t', index=False)
